chore: Move `job_queue/` to `application/job_queue/`
Also:
 - Move `triton_vm_job_queue/` to `application/triton_vm_job_queue/`.

The job queues have more to do with making the particular application
run smoothly, than they do with a) the particular logic of
synchronizing to the rest of the network, or b) participating in the
blockchain protocol in a conforming manner. As such, it belongs in the
category "application".

Furthermore, a negligible fraction of development time was spent in
these directories, suggesting that they are not good fits for first
class citizenship.

The motivation for keeping `job_queue/` and `triton_vm_job_queue/`
separate is because `job_queue/` makes more abstraction still. It is
quite independent of Triton VM and indeed of the rest of the
application. It is a prime candidate for factoring out into its own
crate.
