+++ neptune-core/src/api/tx_initiation/error.rs
@@ -5,8 +5,8 @@ use tasm_lib::prelude::Digest;
 use crate::api::export::BlockHeight;
 use crate::api::export::NativeCurrencyAmount;
 use crate::api::export::RecordTransactionError;
-use crate::job_queue::errors::AddJobError;
-use crate::job_queue::errors::JobHandleError;
+use crate::application::job_queue::errors::AddJobError;
+use crate::application::job_queue::errors::JobHandleError;
 use crate::models::blockchain::transaction::transaction_proof::TransactionProofType;
 use crate::models::proof_abstractions::tasm::prover_job::ProverJobError;
 use crate::state::tx_proving_capability::TxProvingCapability;
