+++ neptune-core/src/application/mod.rs
@@ -1,2 +1,4 @@
 pub mod config;
 pub mod database;
+pub mod job_queue;
+pub mod triton_vm_job_queue;
