+++ neptune-core/src/models/blockchain/consensus_rule_set.rs
@@ -93,7 +93,7 @@ pub(crate) mod tests {
     use crate::state::wallet::wallet_entropy::WalletEntropy;
     use crate::tests::shared::globalstate::mock_genesis_global_state_with_block;
     use crate::tests::tokio_runtime;
-    use crate::triton_vm_job_queue::vm_job_queue;
+    use crate::application::triton_vm_job_queue::vm_job_queue;
 
     use super::*;
