+++ neptune-core/src/models/blockchain/transaction/validity/tasm/claims/generate_k2o_claim.rs
@@ -139,8 +139,8 @@ mod tests {
 
     use super::*;
     use crate::models::blockchain::transaction::primitive_witness::PrimitiveWitness;
-    use crate::triton_vm_job_queue::TritonVmJobPriority;
-    use crate::triton_vm_job_queue::TritonVmJobQueue;
+    use crate::application::triton_vm_job_queue::TritonVmJobPriority;
+    use crate::application::triton_vm_job_queue::TritonVmJobQueue;
 
     #[test]
     fn unit_test() {
