+++ neptune-core/src/models/blockchain/transaction/validity/tasm/claims/generate_lock_script_claim_template.rs
@@ -91,8 +91,8 @@ mod tests {
     use crate::models::blockchain::transaction::primitive_witness::PrimitiveWitness;
     use crate::models::blockchain::transaction::validity::proof_collection::ProofCollection;
     use crate::models::blockchain::transaction::validity::tasm::claims::new_claim::NewClaim;
-    use crate::triton_vm_job_queue::TritonVmJobPriority;
-    use crate::triton_vm_job_queue::TritonVmJobQueue;
+    use crate::application::triton_vm_job_queue::TritonVmJobPriority;
+    use crate::application::triton_vm_job_queue::TritonVmJobQueue;
 
     impl Function for GenerateLockScriptClaimTemplate {
         fn rust_shadow(
