+++ neptune-core/src/models/blockchain/transaction/validity/tasm/claims/generate_type_script_claim_template.rs
@@ -117,8 +117,8 @@ mod tests {
     use super::*;
     use crate::models::blockchain::transaction::primitive_witness::PrimitiveWitness;
     use crate::models::blockchain::transaction::validity::proof_collection::ProofCollection;
-    use crate::triton_vm_job_queue::TritonVmJobPriority;
-    use crate::triton_vm_job_queue::TritonVmJobQueue;
+    use crate::application::triton_vm_job_queue::TritonVmJobPriority;
+    use crate::application::triton_vm_job_queue::TritonVmJobQueue;
 
     impl Function for GenerateTypeScriptClaimTemplate {
         fn rust_shadow(
